requirements.txt
flask
flask-cors
tensorflow-cpu
opencv-python-headless
numpy
mysql-connector-python
gunicorn